refactor open data publishing, add publish_utils
