+++ open_data/gcs_to_esri.py
@@ -12,50 +12,11 @@ import sys
 
 from loguru import logger
 
-import open_data
+import publish_utils
 from calitp_data_analysis import utils, geography_utils
-from shared_utils import portfolio_utils
-from update_vars import analysis_date
+from update_vars import analysis_date, RUN_ME
 
 catalog = intake.open_catalog("./catalog.yml")
-
-def standardize_column_names(df: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
-    """
-    Standardize how agency is referred to.
-    """
-    RENAME_DICT = {
-        "caltrans_district": "district_name",
-        "organization_source_record_id": "org_id",
-        "organization_name": "org_name"
-    }
-    # these rename hqta datasets
-    # agency_name_primary, agency_name_secondary, etc
-    df.columns = df.columns.str.replace('agency_name', 'agency')
-    
-    df = df.rename(columns = RENAME_DICT)
-    df
-    
-    return df
-
-
-def remove_internal_keys(df: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
-    """
-    Remove columns used in our internal data modeling.
-    Leave only natural identifiers (route_id, shape_id).
-    Remove shape_array_key, gtfs_dataset_key, etc.
-    """
-    exclude_list = [
-        "sec_elapsed", "meters_elapsed", 
-        "name" #schedule_gtfs_dataset_name
-    ]
-    cols = [c for c in df.columns]
-    
-    internal_cols = [c for c in cols if "_key" in c or c in exclude_list] 
-    
-    print(f"drop: {internal_cols}")
-    
-    return df.drop(columns = internal_cols)
-    
 
 def print_info(gdf: gpd.GeoDataFrame):
     """
@@ -89,13 +50,11 @@ if __name__=="__main__":
     logger.add(sys.stderr, 
                format="{time:YYYY-MM-DD at HH:mm:ss} | {level} | {message}", 
                level="INFO")
-    
-    #datasets = list(dict(catalog).keys())
-    datasets = open_data.RUN_ME 
-    
-    for d in datasets :
+        
+    for d in RUN_ME :
         gdf = catalog[d].read().to_crs(geography_utils.WGS84)
-        gdf = standardize_column_names(gdf).pipe(remove_internal_keys)
+        gdf = publish_utils.standardize_column_names(gdf).pipe(
+            publish_utils.remove_internal_keys)
 
         logger.info(f"********* {d} *************")
         print_info(gdf)
