+++ open_data/open_data.py
@@ -4,16 +4,8 @@ Track the metadata updates for all open data portal datasets.
 from pathlib import Path
 
 import metadata_update_pro
-from update_vars import XML_FOLDER, META_JSON
+from update_vars import XML_FOLDER, META_JSON, RUN_ME
 
-RUN_ME = [
-    "ca_hq_transit_areas", 
-    "ca_hq_transit_stops",
-    "ca_transit_routes", 
-    "ca_transit_stops",
-    "speeds_by_stop_segments", 
-    "speeds_by_route_time_of_day",
-]
 
 if __name__=="__main__":
     assert str(Path.cwd()).endswith("open_data"), "this script must be run from open_data directory!"
