+++ open_data/update_vars.py
@@ -14,3 +14,12 @@ XML_FOLDER = Path("xml")
 DEFAULT_XML_TEMPLATE = XML_FOLDER.joinpath(Path("default_pro.xml"))
 META_JSON = Path("metadata.json")
 DATA_DICT_YML = Path("data_dictionary.yml")
+
+RUN_ME = [
+    "ca_hq_transit_areas", 
+    "ca_hq_transit_stops",
+    "ca_transit_routes", 
+    "ca_transit_stops",
+    "speeds_by_stop_segments", 
+    "speeds_by_route_time_of_day",
+]
